Time Service (12-hour formatting) for ALL pages
