+++ views/tutorCancelled.ejs
@@ -27,23 +27,12 @@
                   <p><%= appointment.course %></p>
                   <!--used toUTCString() instead of .toDateString() to get the proper date and slice to look nice-->
                   <p><%= appointment.appointmentDate.toUTCString().slice(0, 16) %></p>
+                  
                   <!--Display in standard time-->
-                  <% 
-                    const [startHour, startMinute] = appointment.startTime.split(":");
-                    const [endHour, endMinute] = appointment.endTime.split(":");
-
-                    let startHourNum = parseInt(startHour);
-                    let endHourNum = parseInt(endHour);
-
-                    const startAMPM = startHourNum >= 12 ? "PM" : "AM";
-                    const endAMPM = endHourNum >= 12 ? "PM" : "AM";
-
-                    startHourNum = startHourNum % 12 || 12;
-                    endHourNum = endHourNum % 12 || 12;
-                  %> 
                   <p><b>
-                    <%= startHourNum + ":" + startMinute + " " + startAMPM %> - <%= endHourNum + ":" + endMinute + " " + endAMPM %>
+                    <%= formatTo12Hour(appointment.startTime) %> - <%= formatTo12Hour(appointment.endTime) %>
                   </b></p>
+
                   <br />
                 </div> <!-- end of tutor shift info div -->
               <% }) %>
